+++ humanlayer-ts/src/index.ts
@@ -1,3 +1,4 @@
+import { HumanLayer, ApprovalMethod, humanlayer } from './approval'
 export * from './models'
-export * from './approval'
 export * from './cloud'
+export { humanlayer, ApprovalMethod, HumanLayer }